Android patch: Soft removal of Calendar.BASE_FIELD_COUNT
The field has been deprecated since API level 26.
This CL should remove the field from the SDK, and
we expect no one uses the field in the runtime, as
javac normally inlines the value in the build time
before this change.

ICU 74 has updated the value, and thus the value needs to be removed
from the Android SDK.

Test: n/a
Change-Id: I2abd72b7bd12fe29acf8ae4f8c9c9105d19140e1